+++ src/azure-firewall/HISTORY.rst
@@ -5,8 +5,8 @@ Release History
 
 2.2.0
 ++++++
-* `az network firewall policy create` : Add parameter `--identities` to support multiple MSIs
-* `az network firewall policy update` : Add parameter `--identities` to support multiple MSIs
+* `az network firewall policy create` : Updated parameter `--identity` to support multiple MSIs
+* `az network firewall policy update` : Updated parameter `--identity` to support multiple MSIs
 
 2.1.1
 ++++++